separated stack and bottom tab navigation, added bottom tab navbar
