--- app/screens/Home.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native'
-
-const Home = () => {
-    return (
-        <View>
-            <Text>Home Screen</Text>
-        </View>
-    )
-}
-
-export default Home;
-
-const styles = StyleSheet.create({})
