--- app/screens/Listing.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native'
-
-const Listing = () => {
-    return (
-        <View>
-            <Text>Listing Screen</Text>
-        </View>
-    )
-}
-
-export default Listing;
-
-const styles = StyleSheet.create({})
