--- app/screens/Profile.js
@@ -1,52 +1,0 @@
-import React from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native'
-import { auth } from '../../Firebase/firebase';
-import { signOut } from 'firebase/auth';
-import { useNavigation } from '@react-navigation/native';
-
-const Profile = () => {
-    const navigation = useNavigation()
-
-    const handleSignOut = () => {
-        signOut(auth)
-        .then (() => {
-            navigation.replace("Login")
-        })
-        .catch(error => alert(error.message))
-    }
-
-    return (
-        <View style={styles.container}>
-            <Text>Email: {auth.currentUser?.email}</Text>
-            <TouchableOpacity
-                onPress={handleSignOut}
-                style={styles.button}
-            >
-                <Text style={styles.buttonText}>Sign out</Text>
-            </TouchableOpacity>
-        </View>
-    )
-}
-
-export default Profile;
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center'
-    },
-    button: {
-        backgroundColor: '#3f9eeb',
-        width: '60%',
-        padding: 15,
-        borderRadius: 10,
-        alignItems: 'center',
-        marginTop: 40,
-    },
-    buttonText: {
-        color: 'white',
-        fontWeight: '700',
-        fontSize: 16,
-    },
-})
