--- app/screens/Search.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Search(props) {
-    return (
-        <div>
-            
-        </div>
-    );
-}
-
-export default Search;
