+++ app/screens/Tabs/Cart.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+
+const Cart = () => {
+  return (
+    <View>
+      <Text>Shopping Cart</Text>
+    </View>
+  );
+};
+
+export default Cart;
