+++ app/screens/Tabs/Home.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { StyleSheet, Text, SafeAreaView } from "react-native";
+
+const Home = () => {
+  return (
+    <SafeAreaView>
+      <Text>Home Screen</Text>
+    </SafeAreaView>
+  );
+};
+
+export default Home;
+
+const styles = StyleSheet.create({});
