+++ app/screens/Tabs/Wishlist.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+
+const Wishlist = () => {
+  return (
+    <View>
+      <Text>Wishlist</Text>
+    </View>
+  );
+};
+
+export default Wishlist;
