+++ package.json
@@ -14,6 +14,7 @@
     "@react-native-firebase/app": "^17.4.0",
     "@react-native-firebase/auth": "17.4.0",
     "@react-native-firebase/database": "17.4.0",
+    "@react-navigation/bottom-tabs": "^6.5.14",
     "@react-navigation/native": "^6.1.10",
     "@react-navigation/native-stack": "^6.9.18",
     "expo": "~50.0.6",
@@ -24,6 +25,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.4",
+    "react-native-gesture-handler": "~2.14.0",
+    "react-native-ionicons": "^4.6.5",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
     "react-native-web": "~0.19.6"
